+++ src/main/kotlin/one/pkg/modpublish/ui/AddDependencyDialog.kt
@@ -22,7 +22,7 @@ import com.intellij.util.ui.JBUI
 import one.pkg.modpublish.data.internal.ModInfo
 import one.pkg.modpublish.data.internal.ModInfos
 import one.pkg.modpublish.data.internal.Selector
-import one.pkg.modpublish.data.internal.TargetType
+import one.pkg.modpublish.data.internal.PublishTarget
 import one.pkg.modpublish.data.local.DependencyInfo
 import one.pkg.modpublish.data.local.DependencyType
 import one.pkg.modpublish.ui.base.BaseDialogWrapper
@@ -110,13 +110,13 @@ class AddDependencyDialog(
             if (parts.size != 2) return ModInfos(ModInfo.of("Invalid project ID format"), null)
 
             val modrinthInfo = if (selector.modrinth && parts[0].isNotBlank()) {
-                TargetType.Modrinth.api.getModInfo(parts[0], project!!).also {
+                PublishTarget.Modrinth.api.getModInfo(parts[0], project!!).also {
                     if (it.failed != null) return ModInfos(it, null)
                 }
             } else null
 
             val curseforgeInfo = if (selector.curseForge && parts[1].isNotBlank()) {
-                TargetType.CurseForge.api.getModInfo(parts[1], project!!).also {
+                PublishTarget.CurseForge.api.getModInfo(parts[1], project!!).also {
                     if (it.failed != null) return ModInfos(null, it)
                 }
             } else null
@@ -124,14 +124,14 @@ class AddDependencyDialog(
             return ModInfos(modrinthInfo, curseforgeInfo)
         } else {
             val modrinthInfo = if (selector.modrinth) {
-                TargetType.Modrinth.api.apply { if (getAB()) updateAB() }
+                PublishTarget.Modrinth.api.apply { if (getAB()) updateAB() }
                     .getModInfo(projectId, project!!).also {
                         if (it.failed != null) return ModInfos(it, null)
                     }
             } else null
 
             val curseforgeInfo = if (selector.curseForge) {
-                TargetType.CurseForge.api.getModInfo(projectId, project!!).also {
+                PublishTarget.CurseForge.api.getModInfo(projectId, project!!).also {
                     if (it.failed != null) return ModInfos(null, it)
                 }
             } else null